"Refactor(package, class): Change class and package name"
- change name of package and class of APIBuilder to APIService
